# backend/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/body-logs": {
      "put": {
        "operationId": "BodyLogsController_upsert_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertBodyLogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BodyLog"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "BodyLogs"
        ]
      },
      "get": {
        "operationId": "BodyLogsController_list_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 200,
              "default": 60,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "measuredOn >= from（ISO date）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "measuredOn <= to（ISO date）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BodyLogCursorPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "BodyLogs"
        ]
      }
    },
    "/api/v1/body-logs/{id}": {
      "delete": {
        "operationId": "BodyLogsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "BodyLogs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stopwatch API",
    "description": "Stopwatch 后端 API（阶段 1-5）。给 Flutter 前端对接。",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpsertBodyLogDto": {
        "type": "object",
        "properties": {
          "measuredOn": {
            "type": "string",
            "example": "2026-09-01",
            "description": "ISO date"
          },
          "weightKg": {
            "type": "number",
            "example": 72.4
          },
          "bodyFatPct": {
            "type": "number",
            "example": 18.2
          },
          "waistCm": {
            "type": "number",
            "example": 80
          },
          "armCm": {
            "type": "number",
            "example": 40
          },
          "thighCm": {
            "type": "number",
            "example": 58
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "measuredOn"
        ]
      },
      "BodyLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "number"
          },
          "measuredOn": {
            "type": "string",
            "format": "date",
            "example": "2026-09-01"
          },
          "weightKg": {
            "type": "number",
            "example": 72.4
          },
          "bodyFatPct": {
            "type": "number",
            "example": 18.2,
            "description": "体脂率 %"
          },
          "waistCm": {
            "type": "number",
            "example": 80,
            "description": "腰围 cm"
          },
          "armCm": {
            "type": "number",
            "example": 40,
            "description": "臂围 cm"
          },
          "thighCm": {
            "type": "number",
            "example": 58,
            "description": "大腿围 cm"
          },
          "note": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "measuredOn",
          "createdAt",
          "updatedAt"
        ]
      },
      "BodyLogCursorPage": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BodyLog"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "下一页游标；null = 没有更多"
          }
        },
        "required": [
          "data",
          "nextCursor"
        ]
      },
...
    }
  }
}